web_server/requirements.txt
flask
flask-cors
groq
python-dotenv
requests
pyopenssl
